Wrap ktlint operations/rules in mockkable objects
